backend/requirements.txt
# 后端依赖 - FastAPI 架构
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
pydantic>=2.5.0
python-multipart>=0.0.6
email-validator>=2.0.0

# 基础依赖
bcrypt>=3.2.0
cryptography>=3.4.0
beautifulsoup4>=4.9.0
python-dotenv>=1.0.0
httpx>=0.25.0

# 保留 pymysql 用于迁移期间的兼容性（可选）
# pymysql>=1.0.0

